Add `From<DOT>` for `Graph` and `DiGraph`
