+++ Cargo.toml
@@ -2,15 +2,15 @@
 name = "causal-hub"
 version = "0.1.1"
 edition = "2021"
-authors = [ "Alessio Zanga <alessio.zanga@outlook.it>" ]
+authors = ["Alessio Zanga <alessio.zanga@outlook.it>"]
 description = "A hub for Causal Data Science."
 readme = "README.md"
 license = "Apache-2.0 OR MIT"
 homepage = "https://github.com/AlessioZanga/causal-hub"
 repository = "https://github.com/AlessioZanga/causal-hub"
 documentation = "https://docs.rs/causal-hub"
-keywords = [ "causality", "causal-inference", "causal-discovery" ]
-categories = [ "algorithms", "data-structures", "mathematics", "science" ]
+keywords = ["causality", "causal-inference", "causal-discovery"]
+categories = ["algorithms", "data-structures", "mathematics", "science"]
 
 [dependencies]
 approx = "^0.5"
@@ -47,4 +47,4 @@ serde_json = "^1.0"
 test-log = "^0.2"
 
 [package.metadata.docs.rs]
-rustdoc-args = [ "--html-in-header", "./docs/header.html" ]
+rustdoc-args = ["--html-in-header", "./docs/header.html"]
